release: Update OpenAPI spec to ECE 2.12.0 (MS62) (#352)
Updates the OpenAPI spec to match the upcoming `ECE 2.12.0` version and
all the generated Go code.

+++ pkg/models/dts_dimension_costs.go
@@ -38,7 +38,7 @@ type DtsDimensionCosts struct {
 
 	// Costs associated to the Data Transfer and Storage (DTS) dimensions for an organization
 	// Required: true
-	Cost *int64 `json:"cost"`
+	Cost *float64 `json:"cost"`
 
 	// DTS dimension name
 	// Required: true
